feat(kernel)!: refuse out-of-range signal geometry at entry, typed (J.1+E.7+E.8) (#222)
* feat(kernel)!: refuse out-of-range signal geometry at entry, typed

BREAKING on the JSON wire. Both parse routes previously normalized
silently - start bits and bit lengths were mod-reduced into range at
the boundary, and big-endian signals could be relocated by the
conversion's floor - so format_dbc_text's guarantee was measured
against values the caller never sent, and the error class of the
refusals that did fire depended on where the reduction landed.

The kernel now refuses. A shared geometry decider family (one Dec form
per condition, parameterized by the containing message's frame
capacity - the DLC already distinguishes classic CAN from CAN-FD, so
the bound is the frame's, not a global constant) gates BOTH routes
with typed errors naming the submitted values; the silent clamps and
the monus-reachable relocation retire. The big-endian no-wrap check is
re-aimed at the pre-conversion MSB position: textbook Motorola layouts
the JSON route falsely refused now load on both routes, restoring the
kernel's closure under its own emission (regression-tested). A
present-but-non-natural numeric field now gets its own truthful error
instead of being misreported as missing.

The validator's geometry arms consume the same deciders, and a
machine-checked theorem proves gate acceptance implies those arms are
empty - deadness proven, not observed (the arms stay as defense in
depth). The retired clamp-identity lemmas leave the proof tree, and
the previously surplus round-trip hypotheses are now genuinely
consumed by the re-aimed gate. Every binding's response decoder
accepts full-frame CAN-FD signal lengths (the old classic-CAN cap
crashed decoding the kernel's own success response); wire codes follow
the addition/removal protocol with completeness gates updated in
lockstep.

Closes DEFERRED_ITEMS J.1, E.7, and E.8.


* test(go): guard the non-natural-field constant's wire string

Review fix: the exported-constants regression guard covers the other
geometry codes this change introduces but missed the non-natural-field
constant; its row is added so a typo or rename in the wire string
fails the guard like its siblings.


---------

+++ cpp/include/aletheia/types.hpp
@@ -233,7 +233,7 @@ static_assert(!detail::StrongOfable<RationalFactor, double, int>,
 // ---------------------------------------------------------------------------
 
 using BitPosition = Strong<struct BitPositionTag, std::uint16_t>;
-using BitLength = Strong<struct BitLengthTag, std::uint8_t>;
+using BitLength = Strong<struct BitLengthTag, std::uint16_t>;
 using PropertyIndex = Strong<struct PropertyIndexTag, std::size_t>;
 using MultiplexValue = Strong<struct MultiplexValueTag, std::uint32_t>;
